+++ pubnub-kotlin/pubnub-kotlin-impl/src/test/kotlin/com/pubnub/api/legacy/endpoints/push/AddChannelsToPushTest.kt
@@ -51,28 +51,7 @@ class AddChannelsToPushTest : BaseTest() {
         val requests = WireMock.findAll(WireMock.getRequestedFor(WireMock.urlMatching("/.*")))
         assertEquals(1, requests.size)
         assertEquals("ch1,ch2,ch3", requests[0].queryParameter("add").firstValue())
-        assertEquals("gcm", requests[0].queryParameter("type").firstValue())
-        assertFalse(requests[0].queryParameter("environment").isPresent)
-        assertFalse(requests[0].queryParameter("topic").isPresent)
-    }
-
-    @Test
-    fun testAddMicrosoftSuccessSync() {
-        WireMock.stubFor(
-            WireMock.get(WireMock.urlPathEqualTo("/v1/push/sub-key/mySubscribeKey/devices/niceDevice"))
-                .willReturn(WireMock.aResponse().withBody("[1, \"Modified Channels\"]")),
-        )
-
-        pubnub.addPushNotificationsOnChannels(
-            deviceId = "niceDevice",
-            pushType = PNPushType.MPNS,
-            channels = listOf("ch1", "ch2", "ch3"),
-        ).sync()
-
-        val requests = WireMock.findAll(WireMock.getRequestedFor(WireMock.urlMatching("/.*")))
-        assertEquals(1, requests.size)
-        assertEquals("ch1,ch2,ch3", requests[0].queryParameter("add").firstValue())
-        assertEquals("mpns", requests[0].queryParameter("type").firstValue())
+        assertEquals("fcm", requests[0].queryParameter("type").firstValue())
         assertFalse(requests[0].queryParameter("environment").isPresent)
         assertFalse(requests[0].queryParameter("topic").isPresent)
     }