+++ pubnub-kotlin/pubnub-kotlin-impl/src/test/kotlin/com/pubnub/internal/suite/push/add/AddChannelsToPushV1TestSuite.kt
@@ -38,7 +38,7 @@ class AddChannelsToPushV1TestSuite :
 
     override fun mappingBuilder() =
         get(urlPathEqualTo("/v1/push/sub-key/mySubscribeKey/devices/12345"))
-            .withQueryParam("type", equalTo("gcm"))
+            .withQueryParam("type", equalTo("fcm"))
             .withQueryParam("add", equalTo("ch1,ch2"))
             .withQueryParam("environment", absent())
             .withQueryParam("topic", absent())